requirements.txt
flask==1.1.1
flask_jwt_extended==3.24.1
cryptography==2.8
grpcio==1.28.1
protobuf==3.11.3
